Seperate views to different files
All the views have been seperated accordingly to make the project
more modularized
