+++ project/urls.py
@@ -1,7 +1,7 @@
 from django.contrib import admin
 from django.urls import include, path
 
-from url_shortener.views import RedirectView
+from url_shortener.views.redirect import RedirectView
 
 urlpatterns = [
     path("admin/", admin.site.urls),
